Standardised patching metrics and moved them to utils along with direction cache creator
