+++ utils/circuit_analysis.py
@@ -191,8 +191,8 @@ def get_log_probs(
 def log_prob_diff_noising(
     logits: Float[Tensor, "batch seq d_vocab"],
     answer_tokens: Float[Tensor, "batch n_pairs"],
-    flipped_log_prob: float,
-    clean_log_prob: float,
+    flipped_value: float,
+    clean_value: float,
     return_tensor: bool = False,
 ) -> Float[Tensor, ""]:
     """
@@ -200,17 +200,18 @@ def log_prob_diff_noising(
     same as on clean input, and 1 when performance is same as on flipped input.
     """
     log_prob = get_log_probs(logits, answer_tokens)
-    ld = ((log_prob - clean_log_prob) / (flipped_log_prob  - clean_log_prob))
+    ld = ((log_prob - clean_value) / (flipped_value  - clean_value))
     if return_tensor:
         return ld
     else:
         return ld.item()
     
+
 def log_prob_diff_denoising(
     logits: Float[Tensor, "batch seq d_vocab"],
     answer_tokens: Float[Tensor, "batch n_pairs 2"],
-    flipped_log_prob: float,
-    clean_log_prob: float,
+    flipped_value: float,
+    clean_value: float,
     return_tensor: bool = False,
 ) -> Float[Tensor, ""]:
     """
@@ -218,17 +219,18 @@ def log_prob_diff_denoising(
     same as on flipped input, and 1 when performance is same as on clean input.
     """
     log_prob = get_log_probs(logits, answer_tokens)
-    ld = ((log_prob - flipped_log_prob) / (clean_log_prob  - flipped_log_prob))
+    ld = ((log_prob - flipped_value) / (clean_value  - flipped_value))
     if return_tensor:
         return ld
     else:
         return ld.item()
 
+
 def logit_diff_denoising(
     logits: Float[Tensor, "batch seq d_vocab"],
     answer_tokens: Float[Tensor, "batch n_pairs 2"],
-    flipped_logit_diff: float,
-    clean_logit_diff: float,
+    flipped_value: float,
+    clean_value: float,
     return_tensor: bool = False,
 ) -> Float[Tensor, ""]:
     '''
@@ -236,17 +238,54 @@ def logit_diff_denoising(
     same as on flipped input, and 1 when performance is same as on clean input.
     '''
     patched_logit_diff = get_logit_diff(logits, answer_tokens)
-    ld = ((patched_logit_diff - flipped_logit_diff) / (clean_logit_diff  - flipped_logit_diff))
+    ld = ((patched_logit_diff - flipped_value) / (clean_value  - flipped_value))
     if return_tensor:
         return ld
     else:
         return ld.item()
 
 
+def prob_diff_denoising(
+    logits: Float[Tensor, "batch seq d_vocab"],
+    clean_value: float,
+    corrupted_value: float,
+    answer_tokens: Float[Tensor, "batch 2"],
+    return_tensor: bool = False,
+) -> float:
+    '''
+    Linear function of prob diff, calibrated so that it equals 0 when performance is
+    same as on flipped input, and 1 when performance is same as on clean input.
+    '''
+    
+    patched_logit_diff = get_prob_diff(logits, answer_tokens)
+    ld = ((patched_logit_diff - corrupted_value) / (clean_value  - corrupted_value)).item()
+    if return_tensor:
+        return ld
+    else:
+        return ld.item()
+    
+
+def logit_flip_denoising(
+    logits: Float[Tensor, "batch seq d_vocab"],
+    answer_tokens: Float[Tensor, "batch pair 2"] ,
+    corrupted_value: float,
+    clean_value: float,
+) -> Float[Tensor, ""]:
+    '''
+    Linear function of logit diff, calibrated so that it equals 0 when performance is
+    same as on flipped input, and 1 when performance is same as on clean input.
+    Moves in discrete jumps based on whether logit diffs are closer to clean or corrupted.
+    '''
+    patched_logit_diff = get_logit_diff(logits, answer_tokens, per_prompt=True)
+    clean_distances = (clean_value - patched_logit_diff).abs()
+    corrupt_distances = (corrupted_value - patched_logit_diff).abs()
+    return (clean_distances < corrupt_distances).float().mean().item()
+
+
 def logit_diff_noising(
     logits: Float[Tensor, "batch seq d_vocab"],
-    clean_logit_diff: float,
-    corrupted_logit_diff: float,
+    clean_value: float,
+    corrupted_value: float,
     answer_tokens: Float[Tensor, "batch n_pairs 2"],
     return_tensor: bool = False,
 ) -> float:
@@ -255,12 +294,14 @@ def logit_diff_noising(
     and -1 when performance has been destroyed (i.e. is same as ABC dataset).
     '''
     patched_logit_diff = get_logit_diff(logits, answer_tokens)
-    ld = ((patched_logit_diff - clean_logit_diff) / (clean_logit_diff - corrupted_logit_diff))
+    ld = ((patched_logit_diff - clean_value) / (clean_value - corrupted_value))
 
     if return_tensor:
         return ld
     else:
         return ld.item()
+    
+
 
 
 # =============== LOGIT LENS UTILS ===============
@@ -396,3 +437,72 @@ def get_knockout_perf_drop(model, heads_to_ablate, clean_tokens, metric):
     return ablated_logit_diff
 
 
+def create_cache_for_dir_patching(
+    clean_cache: ActivationCache, 
+    corrupted_cache: ActivationCache, 
+    sentiment_dir: Float[Tensor, "d_model"],
+    model: HookedTransformer,
+) -> ActivationCache:
+    '''
+    We patch the sentiment direction from corrupt to clean
+    '''
+    cache_dict = dict()
+    for act_name, clean_value in clean_cache.items():
+        is_result = act_name.endswith('result')
+        is_resid = (
+            act_name.endswith('resid_pre') or
+            act_name.endswith('resid_post') or
+            act_name.endswith('attn_out') or
+            act_name.endswith('mlp_out')
+        )
+        if is_resid:
+            clean_value = clean_value.to(device)
+            corrupt_value = corrupted_cache[act_name].to(device)
+            corrupt_proj = einops.einsum(
+                corrupt_value, sentiment_dir, 'b s d, d -> b s'
+            )
+            clean_proj = einops.einsum(
+                clean_value, sentiment_dir, 'b s d, d -> b s'
+            )
+            sentiment_dir_broadcast = einops.repeat(
+                sentiment_dir, 'd -> b s d', 
+                b=corrupt_value.shape[0], 
+                s=corrupt_value.shape[1], 
+            )
+            proj_diff = einops.repeat(
+                clean_proj - corrupt_proj, 
+                'b s -> b s d', 
+                d=corrupt_value.shape[-1]
+            )
+            sentiment_adjustment = proj_diff * sentiment_dir_broadcast
+            cache_dict[act_name] = (
+                corrupt_value + sentiment_adjustment
+            )
+        elif is_result:
+            clean_value = clean_value.to(device)
+            corrupt_value = corrupted_cache[act_name].to(device)
+            corrupt_proj = einops.einsum(
+                corrupt_value, sentiment_dir, 'b s h d, d -> b s h'
+            )
+            clean_proj = einops.einsum(
+                clean_value, sentiment_dir, 'b s h d, d -> b s h'
+            )
+            sentiment_dir_broadcast = einops.repeat(
+                sentiment_dir, 'd -> b s h d', 
+                b=corrupt_value.shape[0], 
+                s=corrupt_value.shape[1], 
+                h=corrupt_value.shape[2]
+            )
+            proj_diff = einops.repeat(
+                clean_proj - corrupt_proj, 
+                'b s h -> b s h d', 
+                d=corrupt_value.shape[3]
+            )
+            sentiment_adjustment = proj_diff * sentiment_dir_broadcast
+            cache_dict[act_name] = (
+                corrupt_value + sentiment_adjustment
+            )
+        else:
+            cache_dict[act_name] = clean_value
+
+    return ActivationCache(cache_dict, model)
